+++ gridkit-py/tests/test_gridkit/test_hex_grid.py
@@ -320,16 +320,14 @@ def test_to_shapely_1_cell(shape):
     # The bits outside the rectangle perfectly compensate for the missing bits inside the rectangle
 
 
-@pytest.mark.parametrize("as_mp", [True, False])
 @pytest.mark.parametrize("shape", ["flat", "pointy"])
-def test_to_shapely(as_mp, shape):
+def test_to_shapely(shape):
     grid = HexGrid(size=1.5, shape=shape)
     ids = [[-3, 2], [3, -6]]
-    geoms = grid.to_shapely(ids, as_multipolygon=as_mp)
+    geoms = grid.to_shapely(ids)
 
-    if as_mp:
-        assert isinstance(geoms, shapely.geometry.MultiPolygon)
-        geoms = geoms.geoms
+    assert isinstance(geoms, shapely.geometry.MultiPolygon)
+    geoms = geoms.geoms
 
     centroids = [[geom.centroid.x, geom.centroid.y] for geom in geoms]
     numpy.testing.assert_allclose(centroids, grid.centroid(ids))
@@ -418,12 +416,12 @@ def test_is_aligned_with():
     other_grid = HexGrid(size=1.2, shape="flat")
     is_aligned, reason = grid.is_aligned_with(other_grid)
     assert not is_aligned
-    assert "shape" in reason
+    assert "orientation" in reason
 
     other_grid = HexGrid(size=1.1, shape="flat", offset=(1, 1), crs=4326)
     is_aligned, reason = grid.is_aligned_with(other_grid)
     assert not is_aligned
-    assert all(attr in reason for attr in ["CRS", "shape", "cellsize", "offset"])
+    assert all(attr in reason for attr in ["CRS", "orientation", "cellsize", "offset"])
 
     with pytest.raises(TypeError):
         other_grid = 1